+++ src/tables/network.rs
@@ -1,6 +1,6 @@
 //! The `network` table.
 
-use std::{borrow::Cow, mem::size_of};
+use std::borrow::Cow;
 
 use anyhow::Result;
 use chrono::{DateTime, Utc};
@@ -19,7 +19,6 @@ pub struct Network {
     pub name: String,
     pub description: String,
     pub networks: HostNetworkGroup,
-    pub customer_ids: Vec<u32>,
     tag_ids: Vec<u32>,
     pub creation_time: DateTime<Utc>,
 }
@@ -30,7 +29,6 @@ impl Network {
         name: String,
         description: String,
         networks: HostNetworkGroup,
-        customer_ids: Vec<u32>,
         tag_ids: Vec<u32>,
     ) -> Self {
         Self {
@@ -38,7 +36,6 @@ impl Network {
             name,
             description,
             networks,
-            customer_ids,
             tag_ids: Self::clean_up(tag_ids),
             creation_time: Utc::now(),
         }
@@ -55,12 +52,6 @@ impl Network {
             .map_err(|idx| anyhow::anyhow!("{idx}"))
     }
 
-    fn delete_customer(&mut self, customer_id: u32) -> bool {
-        let prev_len = self.customer_ids.len();
-        self.customer_ids.retain(|&id| id != customer_id);
-        prev_len != self.customer_ids.len()
-    }
-
     fn clean_up(mut tag_ids: Vec<u32>) -> Vec<u32> {
         tag_ids.sort_unstable();
         tag_ids.dedup();
@@ -71,17 +62,12 @@ impl Network {
 impl FromKeyValue for Network {
     fn from_key_value(key: &[u8], value: &[u8]) -> Result<Self> {
         let value: Value = super::deserialize(value)?;
-        let (name, id) = key.split_at(key.len() - size_of::<u32>());
-        let name = std::str::from_utf8(name)?.to_owned();
-        let mut buf = [0; size_of::<u32>()];
-        buf.copy_from_slice(id);
-        let id = u32::from_be_bytes(buf);
+        let name = std::str::from_utf8(key)?.to_owned();
         Ok(Self {
-            id,
+            id: value.id,
             name,
             description: value.description,
             networks: value.networks,
-            customer_ids: value.customer_ids,
             tag_ids: value.tag_ids,
             creation_time: value.creation_time,
         })
@@ -105,28 +91,24 @@ impl Indexable for Network {
         self.id
     }
 
-    fn make_indexed_key(key: Cow<[u8]>, index: u32) -> Cow<[u8]> {
-        let mut key = key.into_owned();
-        key.extend(index.to_be_bytes());
-
-        Cow::Owned(key)
+    /// Returns the key based solely on the name. This enforces global uniqueness by name.
+    fn make_indexed_key(key: Cow<[u8]>, _index: u32) -> Cow<[u8]> {
+        key
     }
 
     fn indexed_key(&self) -> Cow<'_, [u8]> {
-        let mut key = self.name.as_bytes().to_vec();
-        key.extend(self.id.to_be_bytes());
-        std::borrow::Cow::Owned(key)
+        Cow::Borrowed(self.name.as_bytes())
     }
 
     fn value(&self) -> Vec<u8> {
         let value = Value {
+            id: self.id,
             description: self.description.clone(),
             networks: self.networks.clone(),
-            customer_ids: self.customer_ids.clone(),
             tag_ids: self.tag_ids.clone(),
             creation_time: self.creation_time,
         };
-        super::serialize(&value).expect("deserialization error")
+        super::serialize(&value).expect("serialization error")
     }
 
     fn set_index(&mut self, index: u32) {
@@ -136,9 +118,9 @@ impl Indexable for Network {
 
 #[derive(Deserialize, Serialize)]
 struct Value {
+    id: u32,
     description: String,
     networks: HostNetworkGroup,
-    customer_ids: Vec<u32>,
     tag_ids: Vec<u32>,
     creation_time: DateTime<Utc>,
 }
@@ -175,25 +157,9 @@ impl<'d> IndexedTable<'d, Network> {
             let mut network = entry?;
             if let Ok(idx) = network.contains_tag(tag_id) {
                 network.tag_ids.remove(idx);
-                let old = Update::new(None, None, None, None, None);
-                let new = Update::new(None, None, None, None, Some(network.tag_ids));
+                let old = Update::new(None, None, None, None);
+                let new = Update::new(None, None, None, Some(network.tag_ids));
                 self.indexed_map.update(network.id, &old, &new)?;
-            }
-        }
-        Ok(())
-    }
-
-    /// Removes `customer` in all the related entries
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails.
-    pub fn remove_customer(&self, customer: u32) -> Result<()> {
-        let iter = self.iter(Direction::Forward, None);
-        for entry in iter {
-            let mut network = entry?;
-            if network.delete_customer(customer) {
-                self.indexed_map.overwrite(&network)?;
             }
         }
         Ok(())
@@ -213,7 +179,6 @@ pub struct Update {
     name: Option<String>,
     description: Option<String>,
     networks: Option<HostNetworkGroup>,
-    customer_ids: Option<Vec<u32>>,
     tag_ids: Option<Vec<u32>>,
 }
 
@@ -223,7 +188,6 @@ impl Update {
         name: Option<String>,
         description: Option<String>,
         networks: Option<HostNetworkGroup>,
-        customer_ids: Option<Vec<u32>>,
         tag_ids: Option<Vec<u32>>,
     ) -> Self {
         let tag_ids = tag_ids.map(Network::clean_up);
@@ -231,7 +195,6 @@ impl Update {
             name,
             description,
             networks,
-            customer_ids,
             tag_ids,
         }
     }
@@ -259,10 +222,6 @@ impl IndexedMapUpdate for Update {
             value.networks = v.clone();
         }
 
-        if let Some(v) = self.customer_ids.as_deref() {
-            value.customer_ids = v.to_vec();
-        }
-
         if let Some(tag_ids) = self.tag_ids.as_deref() {
             value.tag_ids = Network::clean_up(tag_ids.to_vec());
         }
@@ -283,11 +242,6 @@ impl IndexedMapUpdate for Update {
         }
         if let Some(v) = &self.networks
             && v != &value.networks
-        {
-            return false;
-        }
-        if let Some(v) = self.customer_ids.as_deref()
-            && v != value.customer_ids
         {
             return false;
         }
@@ -314,19 +268,15 @@ mod test {
         let (_permit, store) = setup_store();
         let table = store.network_map();
 
-        let mut network = create_network("TestNetwork", "TestDescription", vec![1, 2], vec![1, 2]);
+        let mut network = create_network("TestNetwork", "TestDescription", vec![1, 2]);
         let inserted_id1 = table.insert(network.clone()).unwrap();
         network.id = inserted_id1;
 
         let retrieved_network = table.get_by_id(inserted_id1).unwrap().unwrap();
         assert_eq!(retrieved_network, network);
 
-        let inserted_id2 = table.insert(network.clone()).unwrap();
-        network.id = inserted_id2;
-        let retrieved_network = table.get_by_id(inserted_id2).unwrap().unwrap();
-        assert_eq!(retrieved_network, network);
-
-        assert!(inserted_id1 != inserted_id2);
+        // Inserting the same network name again should fail (global uniqueness)
+        assert!(table.insert(network.clone()).is_err());
     }
 
     #[test]
@@ -334,9 +284,9 @@ mod test {
         let (_permit, store) = setup_store();
         let table = store.network_map();
 
-        let mut network1 = create_network("Network1", "Description1", vec![1], vec![1, 2]);
-        let mut network2 = create_network("Network2", "Description2", vec![2], vec![2]);
-        let mut network3 = create_network("Network3", "Description3", vec![3], vec![3, 1]);
+        let mut network1 = create_network("Network1", "Description1", vec![1, 2]);
+        let mut network2 = create_network("Network2", "Description2", vec![2]);
+        let mut network3 = create_network("Network3", "Description3", vec![3, 1]);
 
         network1.id = table.insert(network1.clone()).unwrap();
         network2.id = table.insert(network2.clone()).unwrap();
@@ -363,9 +313,9 @@ mod test {
         let (_permit, store) = setup_store();
         let table = store.network_map();
 
-        let mut network1 = create_network("Network1", "Description1", vec![1], vec![1, 2]);
-        let mut network2 = create_network("Network2", "Description2", vec![2], vec![2]);
-        let mut network3 = create_network("Network3", "Description3", vec![3], vec![3, 1]);
+        let mut network1 = create_network("Network1", "Description1", vec![1, 2]);
+        let mut network2 = create_network("Network2", "Description2", vec![2]);
+        let mut network3 = create_network("Network3", "Description3", vec![3, 1]);
 
         network1.id = table.insert(network1.clone()).unwrap();
         network2.id = table.insert(network2.clone()).unwrap();
@@ -374,45 +324,16 @@ mod test {
         table.remove_tag(2).unwrap();
 
         assert_eq!(
-            table.get_by_id(network1.id).unwrap().unwrap().tag_ids,
-            vec![1]
+            table.get_by_id(network1.id).unwrap().unwrap().tag_ids(),
+            &[1]
         );
         assert_eq!(
-            table.get_by_id(network2.id).unwrap().unwrap().tag_ids,
-            Vec::<u32>::new()
+            table.get_by_id(network2.id).unwrap().unwrap().tag_ids(),
+            &[] as &[u32]
         );
         assert_eq!(
-            table.get_by_id(network3.id).unwrap().unwrap().tag_ids,
-            vec![1, 3]
-        );
-    }
-
-    #[test]
-    fn remove_customer() {
-        let (_permit, store) = setup_store();
-        let table = store.network_map();
-
-        let mut network1 = create_network("Network1", "Description1", vec![3, 1], vec![1, 2]);
-        let mut network2 = create_network("Network2", "Description2", vec![2, 1, 4], vec![1, 2]);
-        let mut network3 = create_network("Network3", "Description3", vec![1], vec![1, 2]);
-
-        network1.id = table.insert(network1.clone()).unwrap();
-        network2.id = table.insert(network2.clone()).unwrap();
-        network3.id = table.insert(network3.clone()).unwrap();
-
-        table.remove_customer(2).unwrap();
-
-        assert_eq!(
-            table.get_by_id(network1.id).unwrap().unwrap().customer_ids,
-            vec![3, 1]
-        );
-        assert_eq!(
-            table.get_by_id(network2.id).unwrap().unwrap().customer_ids,
-            vec![1, 4]
-        );
-        assert_eq!(
-            table.get_by_id(network3.id).unwrap().unwrap().customer_ids,
-            vec![1]
+            table.get_by_id(network3.id).unwrap().unwrap().tag_ids(),
+            &[1, 3]
         );
     }
 
@@ -421,25 +342,22 @@ mod test {
         let (_permit, store) = setup_store();
         let mut table = store.network_map();
 
-        let mut network = create_network("Network1", "Description1", vec![1], vec![1, 2]);
+        let mut network = create_network("Network1", "Description1", vec![1, 2]);
         network.id = table.insert(network.clone()).unwrap();
         let old = super::Update::new(
             Some(network.name.clone()),
             Some(network.description.clone()),
             Some(network.networks.clone()),
-            Some(network.customer_ids.clone()),
-            Some(network.tag_ids.clone()),
+            Some(network.tag_ids().to_vec()),
         );
 
-        let mut updated_network =
-            create_network("UpdatedNetwork", "UpdatedDescription", vec![2], vec![3]);
+        let mut updated_network = create_network("UpdatedNetwork", "UpdatedDescription", vec![3]);
         updated_network.creation_time = network.creation_time;
         let update = super::Update::new(
             Some(updated_network.name.clone()),
             Some(updated_network.description.clone()),
             Some(updated_network.networks.clone()),
-            Some(updated_network.customer_ids.clone()),
-            Some(updated_network.tag_ids.clone()),
+            Some(updated_network.tag_ids().to_vec()),
         );
 
         table.update(network.id, &old, &update).unwrap();
@@ -459,17 +377,11 @@ mod test {
         (permit, store)
     }
 
-    fn create_network(
-        name: &str,
-        description: &str,
-        customer_ids: Vec<u32>,
-        tag_ids: Vec<u32>,
-    ) -> Network {
+    fn create_network(name: &str, description: &str, tag_ids: Vec<u32>) -> Network {
         Network::new(
             name.to_string(),
             description.to_string(),
             HostNetworkGroup::default(),
-            customer_ids,
             tag_ids,
         )
     }